refactor: refactor module structures
