+++ src/blender-extension/operators/__init__.py
@@ -1,0 +1,12 @@
+# ------------------------------------------------------------------------------------------
+#  Copyright (c) Natsuneko. All rights reserved.
+#  Licensed under the MIT License. See LICENSE in the project root for license information.
+# ------------------------------------------------------------------------------------------
+
+# pyright: reportUnusedImport=false
+
+from . import fbx
+from . import glb
+from . import obj
+from . import obj_legacy
+from . import vrm
